+++ android/src/main/java/com/hoxfon/react/RNTwilioVoice/TwilioVoiceModule.java
@@ -942,7 +942,7 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
 //            Snackbar.make(coordinatorLayout,
 //                    "Microphone permissions needed. Please allow in your application settings.",
 //                    SNACKBAR_DURATION).show();
-        } else {    
+        } else {
             ActivityCompat.requestPermissions(getCurrentActivity(), new String[]{Manifest.permission.RECORD_AUDIO}, MIC_PERMISSION_REQUEST_CODE);
         }
     }