+++ index.js
@@ -79,7 +79,7 @@ const Twilio = {
     setMuted: TwilioVoice.setMuted,
     setSpeakerPhone: TwilioVoice.setSpeakerPhone,
     sendDigits: TwilioVoice.sendDigits,
-    hold: TwilioVoice.hold,
+    hold: TwilioVoice.setOnHold,
     requestPermissions(senderId) {
         if (Platform.OS === ANDROID) {
             TwilioVoice.requestPermissions(senderId)